adding unit tests and reformatting code

+++ pkg/controller/watermarkpodautoscaler/watermarkpodautoscaler_controller_test.go
@@ -1055,7 +1055,7 @@ func TestCalculateScaleDownLimit(t *testing.T) {
 	}
 }
 
-func makeWPAScaleFactor(scaleUpLimit, scaleDownLimit float64) *v1alpha1.WatermarkPodAutoscaler {
+func makeWPAScaleFactor(scaleUpLimit, scaleDownLimit int32) *v1alpha1.WatermarkPodAutoscaler {
 	return &v1alpha1.WatermarkPodAutoscaler{
 		Spec: v1alpha1.WatermarkPodAutoscalerSpec{
 			ScaleDownLimitFactor: scaleDownLimit,
